Refactor protocol methods to use specific type enums instead of uint32
- Updated `SetContentType` in `content_type_v1` to accept `WpContentTypeV1Type`.
- Modified `SetShape` in `cursor_shape_v1` to use `WpCursorShapeDeviceV1Shape`.
- Changed `Flags` in `ExtBackgroundEffectManagerV1CapabilitiesEvent` to `ExtBackgroundEffectManagerV1Capability`.
- Updated `CreateSession` in `ext_image_copy_capture_v1` to accept `ExtImageCopyCaptureManagerV1Options`.
- Refactored `Format` in `ExtImageCopyCaptureSessionV1ShmFormatEvent` to `WlShmFormat`.
- Changed `Transform` in `ExtImageCopyCaptureFrameV1TransformEvent` to `WlOutputTransform`.
- Updated `Reason` in `ExtImageCopyCaptureFrameV1FailedEvent` to `ExtImageCopyCaptureFrameV1FailureReason`.
- Modified `Capabilities` in `ExtWorkspaceGroupHandleV1CapabilitiesEvent` to `ExtWorkspaceGroupHandleV1GroupCapabilities`.
- Changed `State` in `ExtWorkspaceHandleV1StateEvent` to `ExtWorkspaceHandleV1State`.
- Updated `Capability` in `ZwpFullscreenShellV1CapabilityEvent` to `ZwpFullscreenShellV1Capability`.
- Refactored `Flags` in `ZwpLinuxDmabufFeedbackV1TrancheFlagsEvent` to `ZwpLinuxDmabufFeedbackV1TrancheFlags`.
- Updated `SetPresentationHint` in `tearing_control_v1` to accept `WpTearingControlV1PresentationHint`.
- Changed `SetContentType` in `text_input_unstable_v1` to use `ZwpTextInputV1ContentHint` and `ZwpTextInputV1ContentPurpose`.
- Modified `SetTextChangeCause` in `text_input_unstable_v3` to accept `ZwpTextInputV3ChangeCause`.
- Updated `GetLayerSurface` in `wlr_layer_shell_unstable_v1` to use `ZwlrLayerShellV1Layer`.
- Refactored `SetMode` in `xdg_decoration_unstable_v1` to accept `ZxdgToplevelDecorationV1Mode`.
- Changed `SetAnchor` and `SetGravity` in `xdg_shell` and `xdg_shell_unstable_v6` to use specific enum types.

+++ examples/rainbow/main.go
@@ -43,7 +43,7 @@ func handle(err error) {
 // Event handlers for Display
 
 func (d *Display) handleShmFormat(ev client.WlShmFormatEvent) {
-	if ev.Format == uint32(client.WlShmFormatXrgb8888) {
+	if ev.Format == client.WlShmFormatXrgb8888 {
 		d.hasXrgb = true
 	}
 }
@@ -144,7 +144,7 @@ func createShmBuffer(w *Window, buffer *Buffer, width int, height int, format cl
 		return err
 	}
 
-	buf, err := pool.CreateBuffer(0, int32(width), int32(height), int32(stride), uint32(format))
+	buf, err := pool.CreateBuffer(0, int32(width), int32(height), int32(stride), format)
 	if err != nil {
 		return err
 	}
